Moved PaymentsHub to the payments package
