+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/payments/hub/PaymentsHubTapToPayUnavailableHandlerTest.kt
@@ -1,6 +1,7 @@
-package com.woocommerce.android.ui.payments.cardreader.hub
+package com.woocommerce.android.ui.payments.hub
 
 import com.woocommerce.android.R
+import com.woocommerce.android.ui.payments.hub.PaymentsHubTapToPayUnavailableHandler
 import com.woocommerce.android.ui.payments.taptopay.TapToPayAvailabilityStatus
 import com.woocommerce.android.viewmodel.MultiLiveEvent
 import org.assertj.core.api.Assertions.assertThat
